# docs/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/workouts/analytics/summary": {
      "get": {
        "tags": [
          "WorkoutAnalytics"
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutSummaryDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutSummaryDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutSummaryDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/workouts/analytics/exercise-history": {
      "get": {
        "tags": [
          "WorkoutAnalytics"
        ],
        "parameters": [
          {
            "name": "exerciseId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseHistoryDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseHistoryDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseHistoryDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExerciseHistoryDto": {
        "type": "object",
        "properties": {
          "exerciseId": {
            "type": "integer",
            "format": "int32"
          },
          "exerciseName": {
            "type": "string",
            "nullable": true
          },
          "from": {
            "type": "string",
            "format": "date"
          },
          "to": {
            "type": "string",
            "format": "date"
          },
          "points": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExerciseHistoryPointDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ExerciseHistoryPointDto": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "integer",
            "format": "int32"
          },
          "performedOn": {
            "type": "string",
            "format": "date"
          },
          "setCount": {
            "type": "integer",
            "format": "int32"
          },
          "totalReps": {
            "type": "integer",
            "format": "int32"
          },
          "totalVolume": {
            "type": "number",
            "format": "double"
          },
          "maxWeight": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "bestEstimatedOneRepMax": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "MuscleGroupEnum": {
        "enum": [
          "Other",
          "Chest",
          "Back",
          "Shoulders",
          "Biceps",
          "Triceps",
          "Forearms",
          "Abs",
          "Glutes",
          "Quadriceps",
          "Hamstrings",
          "Calves",
          "FullBody",
          "Cardio"
        ],
        "type": "string"
      },
      "MuscleGroupVolumeDto": {
        "type": "object",
        "properties": {
          "muscleGroup": {
            "allOf": [
              {
                "$ref": "#/components/schemas/MuscleGroupEnum"
              }
            ]
          },
          "setCount": {
            "type": "integer",
            "format": "int32"
          },
          "totalReps": {
            "type": "integer",
            "format": "int32"
          },
          "totalVolume": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
...
      "WorkoutSummaryDto": {
        "type": "object",
        "properties": {
          "from": {
            "type": "string",
            "format": "date"
          },
          "to": {
            "type": "string",
            "format": "date"
          },
          "sessionCount": {
            "type": "integer",
            "format": "int32"
          },
          "exerciseCount": {
            "type": "integer",
            "format": "int32"
          },
          "setCount": {
            "type": "integer",
            "format": "int32"
          },
          "totalReps": {
            "type": "integer",
            "format": "int32"
          },
          "totalVolume": {
            "type": "number",
            "format": "double"
          },
          "totalDurationSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "averageDurationSeconds": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "byMuscleGroup": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MuscleGroupVolumeDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Please enter ONLY token",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      },
      "x-time-zone": {
        "type": "apiKey",
        "description": "Specify the user's timezone in IANA format (e.g., 'America/New_York')",
        "name": "x-time-zone",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    },
    {
      "x-time-zone": [ ]
    }
  ]
}